Created SimpleActionExecutor class
It takes a GerritEventFilter and an ActionForGerritEvent.

It executes the same action for every filtered event.

The GerritClient is initialized here. It would
probably be nice to create a base class for
action executors, so we could have different
kinds of them. Then, initialization of the
client would be part of this base class.