Improve satisfaction_for_annoying_review_requesters
Some users upload changes to gerrit with the list
of reviewers already set.

This does not trigger ReviwerAddedEvent, but a PatchsetCreatedEvent

To capture this situations now we filter PatchsetCreatedEvents
rather than ReviwerAddedEvent.

Ideally, we need a combination of both:
when onwer is annoying, and I have been added as reviewer,
or a patchset is created in a change in which I am already
a reviewer.